Depend on the bucket etag
For the directory sync to take effect on an update, the dynamic
resource needs to actually be updated. The only thing about the
bucket or its content that it depended upon, however, were
properties that don't actually change when the content changes.

The fix is to depend on the bucket's etag. This way anytime the
content etag changes, the syncer will run and expand the new tgz.

Fixes #2.